backend/requirement.txt
Flask==2.3.2
Pillow==10.0.0
pymupdf==1.23.0